Move volumetric to DFT
