+++ tests/atomistics/volumetric/test_volumetric_data.py
@@ -5,7 +5,7 @@
 import unittest
 import numpy as np
 import os
-from pyiron_atomistics.atomistics.volumetric.generic import VolumetricData
+from pyiron_atomistics.dft.volumetric import VolumetricData
 from pyiron_atomistics.atomistics.structure.atoms import Atoms
 from pyiron_atomistics.atomistics.structure.factory import StructureFactory
 from pyiron_atomistics.vasp.volumetric_data import VaspVolumetricData
